apps/preprocessor/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.5.0

# Azure Storage
azure-storage-blob==12.19.0
aiohttp>=3.9.0
aiofiles==23.2.1

# Data processing
pandas==2.1.4
numpy==1.26.4

# HTTP client
httpx==0.25.2

# Scheduling (optional, for schedule mode)
apscheduler==3.10.4

# Logging
python-json-logger==2.0.7

# OpenTelemetry
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0

